aijp

+++ avg.c
@@ -1,6 +1,6 @@
 #include<stdio.h> 
 int main(){
-       double x;
+       long long int x;
        int size = sizeof(x);
        printf("Size of x is: %d", sizeof(x));
        return 0;
